refactor(57): grouped links by category
