+++ src/components/Pages.tsx
@@ -1,0 +1,15 @@
+import { AdminPages } from "../pages/Admin/AdminPages";
+import { FrontPage } from "../pages/MainPage/FrontPage";
+import { PCPages } from "../pages/Products/PC/PCPages";
+import { ShoppingCartPage } from "../pages/ShoppingCart/ShoppingCartPage";
+
+export const Pages = () => {
+    return (
+        <>
+            <FrontPage />
+            <ShoppingCartPage />
+            <PCPages />
+            <AdminPages />
+        </>
+    );
+};
