+++ src/pages/Admin/AdminPages.tsx
@@ -1,0 +1,9 @@
+import { ProcessorBrandPages } from "./ComboData/ProcessorBrandPages";
+
+export const AdminPages = () => {
+    return (
+        <>
+            <ProcessorBrandPages />
+        </>
+    );
+};
